+++ client/src/autoComplete.js
@@ -2,6 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
+// biome-ignore lint/correctness/noUndeclaredDependencies: is defined
 import autoComplete from "@tarekraafat/autocomplete.js";
 
 window.autoComplete = autoComplete;