+++ docs/changes.rst
@@ -18,13 +18,16 @@ Weblate 5.11
 * :ref:`upload` is now tracked in history with details.
 * :ref:`check-c-sharp-format` now supports ``csharp-format`` flag for compatibility with GNU gettext.
 * Changes in string flags are now tracked in history.
+* :doc:`/admin/machine` documentation extended.
 * :ref:`addon-weblate.discovery.discovery` better handles hundredths of matches.
 * Dismissing :ref:`checks` automatically updates propagated strings.
+* Improved rendering of :ref:`additional-flags` and :ref:`additional-explanation` changes in history.
 
 .. rubric:: Bug fixes
 
 * Fixed captcha verification when some time zone was configured.
 * Improved translation propagation performance.
+* Leading and trailing whitespace are now correctly stripped in glossary strings that also contain a :ref:`check-prohibited-initial-character`.
 * Fixed background parsing of newly added translation files.
 
 .. rubric:: Compatibility